backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.15
gunicorn>=21.2.0
whitenoise>=6.5
psycopg2-binary>=2.9   # optional, for Postgres later
mysqlclient
python-dotenv
Pillow>=9.0.0  # Required for ImageField support
